moved test class to the same pkg as the main class
