+++ crud-operations/src/test/java/com/intellectDesign/springBoot/crudoperations/com/intellectDesign/springBoot/crudoperations/controllers/ControllerTest.java
@@ -1,4 +1,4 @@
-package com.intellectDesign.springBoot.crudoperations;
+package com.intellectDesign.springBoot.crudoperations.com.intellectDesign.springBoot.crudoperations.controllers;
 
 import static org.assertj.core.api.Assertions.assertThat;
