+++ crud-operations/src/test/java/com/intellectDesign/springBoot/crudoperations/com/intellectDesign/springBoot/crudoperations/controllers/MockMvcTest.java
@@ -1,4 +1,4 @@
-package com.intellectDesign.springBoot.crudoperations;
+package com.intellectDesign.springBoot.crudoperations.com.intellectDesign.springBoot.crudoperations.controllers;
 
 
 import static org.hamcrest.Matchers.containsString;
